Исправлены ошибки при программных аргументах, класс Archive стал абстрактным. Добавлен код для разархивирования, но пока ещё не проверен в деле. Исправлен путь до файла с логами и улучшено логирование. Добавлено время

+++ HuffmanTree.h
@@ -76,7 +76,7 @@ public:
 		char zeroSymbol = str[0];
 		vector<Node*> nodes;
 
-		for (int i = 0; i < str.size(); ++i)
+		for (unsigned long long i = 0; i < str.size(); ++i)
 		{
 			Node* node = new Node{ str[i] };
 			nodes.push_back(node);
